chore: rename "dummy" data to "sample" data
